Move file helpers to pkg/compliance/utils/file
